+++ fixtures/ext-types/lib/tests/test_generated_bindings.rs
@@ -1,4 +1,5 @@
 uniffi_macros::build_foreign_language_testcases!(
     "tests/bindings/test_imported_types.kts",
     "tests/bindings/test_imported_types.py",
+    "tests/bindings/test_imported_types.swift",
 );