+++ uniffi_bindgen/src/bindings/swift/templates/CallbackInterfaceTemplate.swift
@@ -128,7 +128,7 @@ extension {{ ffi_converter_name }} : FfiConverter {
     // We can use Handle as the FFIType because it's a typealias to UInt64
     typealias FfiType = UniFFICallbackHandle
 
-    static func lift(_ handle: UniFFICallbackHandle) throws -> SwiftType {
+    public static func lift(_ handle: UniFFICallbackHandle) throws -> SwiftType {
         ensureCallbackinitialized();
         guard let callback = handleMap.get(handle: handle) else {
             throw UniffiInternalError.unexpectedStaleHandle
@@ -136,18 +136,18 @@ extension {{ ffi_converter_name }} : FfiConverter {
         return callback
     }
 
-    static func read(from buf: inout (data: Data, offset: Data.Index)) throws -> SwiftType {
+    public static func read(from buf: inout (data: Data, offset: Data.Index)) throws -> SwiftType {
         ensureCallbackinitialized();
         let handle: UniFFICallbackHandle = try readInt(&buf)
         return try lift(handle)
     }
 
-    static func lower(_ v: SwiftType) -> UniFFICallbackHandle {
+    public static func lower(_ v: SwiftType) -> UniFFICallbackHandle {
         ensureCallbackinitialized();
         return handleMap.insert(obj: v)
     }
 
-    static func write(_ v: SwiftType, into buf: inout [UInt8]) {
+    public static func write(_ v: SwiftType, into buf: inout [UInt8]) {
         ensureCallbackinitialized();
         writeInt(&buf, lower(v))
     }