+++ uniffi_bindgen/src/bindings/swift/templates/ErrorTemplate.swift
@@ -15,10 +15,10 @@ public enum {{ type_name }} {
     {%- endif %}
 }
 
-fileprivate struct {{ ffi_converter_name }}: FfiConverterRustBuffer {
+public struct {{ ffi_converter_name }}: FfiConverterRustBuffer {
     typealias SwiftType = {{ type_name }}
 
-    static func read(from buf: inout (data: Data, offset: Data.Index)) throws -> {{ type_name }} {
+    public static func read(from buf: inout (data: Data, offset: Data.Index)) throws -> {{ type_name }} {
         let variant: Int32 = try readInt(&buf)
         switch variant {
 
@@ -46,7 +46,7 @@ fileprivate struct {{ ffi_converter_name }}: FfiConverterRustBuffer {
         }
     }
 
-    static func write(_ value: {{ type_name }}, into buf: inout [UInt8]) {
+    public static func write(_ value: {{ type_name }}, into buf: inout [UInt8]) {
         switch value {
 
         {% if e.is_flat() %}