+++ uniffi_bindgen/src/bindings/swift/templates/Int16Helper.swift
@@ -2,11 +2,11 @@ fileprivate struct FfiConverterInt16: FfiConverterPrimitive {
     typealias FfiType = Int16
     typealias SwiftType = Int16
 
-    static func read(from buf: inout (data: Data, offset: Data.Index)) throws -> Int16 {
+    public static func read(from buf: inout (data: Data, offset: Data.Index)) throws -> Int16 {
         return try lift(readInt(&buf))
     }
 
-    static func write(_ value: Int16, into buf: inout [UInt8]) {
+    public static func write(_ value: Int16, into buf: inout [UInt8]) {
         writeInt(&buf, lower(value))
     }
 }