+++ uniffi_bindgen/src/bindings/swift/templates/OptionalTemplate.swift
@@ -1,7 +1,7 @@
 fileprivate struct {{ ffi_converter_name }}: FfiConverterRustBuffer {
     typealias SwiftType = {{ type_name }}
 
-    static func write(_ value: SwiftType, into buf: inout [UInt8]) {
+    public static func write(_ value: SwiftType, into buf: inout [UInt8]) {
         guard let value = value else {
             writeInt(&buf, Int8(0))
             return
@@ -10,7 +10,7 @@ fileprivate struct {{ ffi_converter_name }}: FfiConverterRustBuffer {
         {{ inner_type|write_fn }}(value, into: &buf)
     }
 
-    static func read(from buf: inout (data: Data, offset: Data.Index)) throws -> SwiftType {
+    public static func read(from buf: inout (data: Data, offset: Data.Index)) throws -> SwiftType {
         switch try readInt(&buf) as Int8 {
         case 0: return nil
         case 1: return try {{ inner_type|read_fn }}(from: &buf)