+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 
 [All changes in [[UnreleasedVersion]]](https://github.com/mozilla/uniffi-rs/compare/v0.21.0...HEAD).
 
+### What's changed
+
+- Added support for Swift external types
+
 ## v0.21.0 - (_2022-10-14_)
 
 [All changes in v0.21.0](https://github.com/mozilla/uniffi-rs/compare/v0.20.0...v0.21.0).